Futile attempt to fix crosshair trailing behind mouse
Apparently roblox made the *actual mouse that is recognized by UI elements and the engine itself* invisible and made an overlay that is on point to your windows mouse.

It looks nicer this way but the reality is that the actual mouse is laggy and trails behind the actual mouse, hence the crosshair trailing behind.